requirements.txt
Django==3.1.7
django-adminlte3==0.1.6
django-seed==0.2.2
psycopg2==2.8.6
python-environ==0.4.54
django-crispy-forms==1.11.2